refactor: ImmutableField.tsをField.tsにリネームし全参照を更新
- 全関連サービスおよびテストでの型・関数参照をField.tsに統一
- 依存関係レポートおよび関連HTMLファイルも更新完了
