+++ app/src/domain/services/ImmutableEliminationService.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { createEmptyField, placePuyoAt } from '../models/ImmutableField'
+import { createEmptyField, placePuyoAt } from '../models/Field.ts'
 import { createPuyo } from '../models/Puyo'
 import {
   eliminateGroups,
